Upload for transfer

+++ src/main/java/com/tiggerbiggo/prima/play/graphics/ColorConvertType.java
@@ -7,9 +7,9 @@ public enum ColorConvertType{
   H(ColorTools::getHue),
   S(ColorTools::getSaturation),
   V(ColorTools::getBrightness),
-  R(c -> (double)c.getRed()/255),
-  G(c -> (double)c.getGreen()/255),
-  B(c -> (double)c.getBlue()/255);
+  R(c -> c.getRed()/255d),
+  G(c -> c.getGreen()/255d),
+  B(c -> c.getBlue()/255d);
 
   Function<Color, Double> convertFunction;
